[feat] added diagonal check and cost multiplier

+++ Dijkstra/AStar.h
@@ -7,8 +7,8 @@
 class AStar
 {
 public:
-	void Algorithm(const std::vector<std::vector<int>>& grid, Vector2 start, Vector2 goal);
-	std::vector<Node*> GetChildrens(Node* current, const std::vector<std::vector<int>>& grid);
+	void Algorithm(const std::vector<std::vector<float>>& grid, Vector2 start, Vector2 goal);
+	std::vector<Node*> GetChildrens(Node* current, const std::vector<std::vector<float>>& grid);
 	float Heuristic(Node* a, Node* b);
 	void CleanupNodes(std::vector<Node*>& nodes);
 };
